move profile E2E tests to test/e2e
