+++ package.json
@@ -64,7 +64,8 @@
 		"vbb-stations-autocomplete": "^4.1.0"
 	},
 	"scripts": {
-		"test": "env NODE_ENV=dev node test/index.js",
-		"prepublishOnly": "npm test | tap-spec"
+		"e2e-test": "env NODE_ENV=dev node test/e2e/index.js | tap-spec",
+		"test": "npm run e2e-test",
+		"prepublishOnly": "npm test"
 	}
 }
