+++ test/e2e/index.js
@@ -1,0 +1,15 @@
+'use strict'
+
+require('./common')
+require('./db')
+require('./vbb')
+require('./bvg')
+require('./oebb')
+require('./insa')
+require('./nahsh')
+require('./cmta')
+require('./nvv')
+require('./sbahn-muenchen')
+require('./saarfahrplan')
+require('./cfl')
+require('./hvv')
