+++ test/e2e/insa.js
@@ -5,18 +5,19 @@ const tape = require('tape')
 const isRoughlyEqual = require('is-roughly-equal')
 
 const {createWhen} = require('./lib/util')
-const createClient = require('..')
-const hvvProfile = require('../p/hvv')
-const products = require('../p/hvv/products')
+const createClient = require('../..')
+const insaProfile = require('../../p/insa')
+const products = require('../../p/insa/products')
 const createValidate = require('./lib/validate-fptf-with')
 const testJourneysStationToStation = require('./lib/journeys-station-to-station')
-const journeysFailsWithNoProduct = require('./lib/journeys-fails-with-no-product')
 const testJourneysStationToAddress = require('./lib/journeys-station-to-address')
 const testJourneysStationToPoi = require('./lib/journeys-station-to-poi')
 const testEarlierLaterJourneys = require('./lib/earlier-later-journeys')
+const journeysFailsWithNoProduct = require('./lib/journeys-fails-with-no-product')
 const testDepartures = require('./lib/departures')
 const testDeparturesInDirection = require('./lib/departures-in-direction')
 const testArrivals = require('./lib/arrivals')
+const testJourneysWithDetour = require('./lib/journeys-with-detour')
 
 const isObj = o => o !== null && 'object' === typeof o && !Array.isArray(o)
 
@@ -24,28 +25,29 @@ const when = createWhen('Europe/Berlin', 'de-DE')
 
 const cfg = {
 	when,
-	// stationCoordsOptional: false,
+	stationCoordsOptional: false,
 	products,
-	// minLatitude: 50.7,
-	// maxLatitude: 53.2,
-	// minLongitude: 10.25,
-	// maxLongitude: 13.4
+	minLatitude: 50.7,
+	maxLatitude: 53.2,
+	minLongitude: 10.25,
+	maxLongitude: 13.4
 }
 
 const validate = createValidate(cfg, {})
 
 const test = tapePromise(tape)
-const client = createClient(hvvProfile, 'public-transport/hafas-client:test')
+const client = createClient(insaProfile, 'public-transport/hafas-client:test')
 
-const tiefstack = '4117'
-const barmbek = '4933'
-const altona = '20626'
-// const hasselbachplatzSternstrasse = '6545'
-// const stendal = '8010334'
-// const dessau = '8010077'
+const magdeburgHbf = '8010224'
+const magdeburgBuckau = '8013456'
+const leiterstr = '7464'
+const hasselbachplatzSternstrasse = '6545'
+const stendal = '8010334'
+const dessau = '8010077'
+const universitaet = '19686'
 
-test('journeys – Hamburg Tiefstack to Hamburg Barmbek', async (t) => {
-	const res = await client.journeys(tiefstack, barmbek, {
+test('journeys – Magdeburg Hbf to Magdeburg-Buckau', async (t) => {
+	const res = await client.journeys(magdeburgHbf, magdeburgBuckau, {
 		results: 4,
 		departure: when,
 		stopovers: true
@@ -55,8 +57,8 @@ test('journeys – Hamburg Tiefstack to Hamburg Barmbek', async (t) => {
 		test: t,
 		res,
 		validate,
-		fromId: tiefstack,
-		toId: barmbek
+		fromId: magdeburgHbf,
+		toId: magdeburgBuckau
 	})
 	t.end()
 })
@@ -67,24 +69,23 @@ test('journeys – fails with no product', (t) => {
 	journeysFailsWithNoProduct({
 		test: t,
 		fetchJourneys: client.journeys,
-		fromId: tiefstack,
-		toId: barmbek,
+		fromId: magdeburgHbf,
+		toId: magdeburgBuckau,
 		when,
 		products
 	})
 	t.end()
 })
 
-test('Hamburg Tiefstack to Gilbertstr. 30, Hamburg', async (t) => {
-	const gilbertstr30 = {
+test('Magdeburg Hbf to 39104 Magdeburg, Sternstr. 10', async (t) => {
+	const sternStr = {
 		type: 'location',
-		id: '970026640',
-		address: 'Hamburg, Gilbertstraße 30',
-		latitude: 53.554791,
-		longitude: 9.95781
+		address: 'Magdeburg - Altenstadt, Sternstraße 10',
+		latitude: 52.118414,
+		longitude: 11.422332
 	}
 
-	const res = await client.journeys(tiefstack, gilbertstr30, {
+	const res = await client.journeys(magdeburgHbf, sternStr, {
 		results: 3,
 		departure: when
 	})
@@ -93,22 +94,22 @@ test('Hamburg Tiefstack to Gilbertstr. 30, Hamburg', async (t) => {
 		test: t,
 		res,
 		validate,
-		fromId: tiefstack,
-		to: gilbertstr30
+		fromId: magdeburgHbf,
+		to: sternStr
 	})
 	t.end()
 })
 
-test('Hamburg Tiefstack to Hamburger Meile', async (t) => {
-	const meile = {
+test('Magdeburg Hbf to Kloster Unser Lieben Frauen', async (t) => {
+	const kloster = {
 		type: 'location',
-		id: '980001825',
+		id: '970012223',
 		poi: true,
-		name: 'Hamburger Meile',
-		latitude: 53.572455,
-		longitude: 10.030541
+		name: 'Magdeburg, Kloster Unser Lieben Frauen (Denkmal)',
+		latitude: 52.127601,
+		longitude: 11.636437
 	}
-	const res = await client.journeys(tiefstack, meile, {
+	const res = await client.journeys(magdeburgHbf, kloster, {
 		results: 3,
 		departure: when
 	})
@@ -117,22 +118,41 @@ test('Hamburg Tiefstack to Hamburger Meile', async (t) => {
 		test: t,
 		res,
 		validate,
-		fromId: tiefstack,
-		to: meile
+		fromId: magdeburgHbf,
+		to: kloster
 	})
 	t.end()
 })
 
-// todo: via works – with detour
-// todo: via works – without detour
+test('journeys: via works – with detour', async (t) => {
+	// Going from Magdeburg, Hasselbachplatz (Sternstr.) (Tram/Bus) to Stendal
+	// via Dessau without detour is currently impossible. We check if the routing
+	// engine computes a detour.
+	const res = await client.journeys(hasselbachplatzSternstrasse, stendal, {
+		via: dessau,
+		results: 1,
+		departure: when,
+		stopovers: true
+	})
+
+	await testJourneysWithDetour({
+		test: t,
+		res,
+		validate,
+		detourIds: [dessau]
+	})
+	t.end()
+})
+
+// todo: without detour
 
 test('earlier/later journeys', async (t) => {
 	await testEarlierLaterJourneys({
 		test: t,
 		fetchJourneys: client.journeys,
 		validate,
-		fromId: tiefstack,
-		toId: barmbek,
+		fromId: magdeburgHbf,
+		toId: magdeburgBuckau,
 		when
 	})
 
@@ -140,7 +160,7 @@ test('earlier/later journeys', async (t) => {
 })
 
 test('trip details', async (t) => {
-	const res = await client.journeys(tiefstack, barmbek, {
+	const res = await client.journeys(magdeburgHbf, magdeburgBuckau, {
 		results: 1, departure: when
 	})
 
@@ -153,8 +173,8 @@ test('trip details', async (t) => {
 	t.end()
 })
 
-test('departures at Hamburg Barmbek', async (t) => {
-	const departures = await client.departures(barmbek, {
+test('departures at Magdeburg Leiterstr.', async (t) => {
+	const departures = await client.departures(leiterstr, {
 		duration: 5, when,
 		stopovers: true
 	})
@@ -163,7 +183,7 @@ test('departures at Hamburg Barmbek', async (t) => {
 		test: t,
 		departures,
 		validate,
-		id: barmbek
+		id: leiterstr
 	})
 	t.end()
 })
@@ -171,8 +191,8 @@ test('departures at Hamburg Barmbek', async (t) => {
 test('departures with station object', async (t) => {
 	const deps = await client.departures({
 		type: 'station',
-		id: tiefstack,
-		name: 'Hamburg Tiefstack',
+		id: magdeburgHbf,
+		name: 'Magdeburg Hbf',
 		location: {
 			type: 'location',
 			latitude: 1.23,
@@ -184,21 +204,21 @@ test('departures with station object', async (t) => {
 	t.end()
 })
 
-test('departures at Barmbek in direction of Altona', async (t) => {
+test('departures at Leiterstr in direction of Universität', async (t) => {
 	await testDeparturesInDirection({
 		test: t,
 		fetchDepartures: client.departures,
 		fetchTrip: client.trip,
-		id: barmbek,
-		directionIds: [altona],
+		id: leiterstr,
+		directionIds: [universitaet],
 		when,
 		validate
 	})
 	t.end()
 })
 
-test('arrivals at Hamburg Barmbek', async (t) => {
-	const arrivals = await client.arrivals(barmbek, {
+test('arrivals at Magdeburg Leiterstr.', async (t) => {
+	const arrivals = await client.arrivals(leiterstr, {
 		duration: 5, when
 	})
 
@@ -206,16 +226,16 @@ test('arrivals at Hamburg Barmbek', async (t) => {
 		test: t,
 		arrivals,
 		validate,
-		id: barmbek
+		id: leiterstr
 	})
 	t.end()
 })
 
 // todo: nearby
 
-test('locations named Elbphilharmonie', async (t) => {
-	const elbphilharmonie = '6242'
-	const locations = await client.locations('Elbphilharmonie', {
+test('locations named Magdeburg', async (t) => {
+	const nordpark = '7480'
+	const locations = await client.locations('nordpark', {
 		results: 20
 	})
 
@@ -225,31 +245,36 @@ test('locations named Elbphilharmonie', async (t) => {
 	t.ok(locations.find(s => s.type === 'stop' || s.type === 'station'))
 	t.ok(locations.find(s => s.poi)) // POIs
 	t.ok(locations.some((l) => {
-		return l.station && l.station.id === elbphilharmonie || l.id === elbphilharmonie
+		return l.station && l.station.id === nordpark || l.id === nordpark
 	}))
 
 	t.end()
 })
 
-test('station Hamburg Barmbek', async (t) => {
-	const s = await client.stop(barmbek)
+test('station Magdeburg-Buckau', async (t) => {
+	const s = await client.stop(magdeburgBuckau)
 
 	validate(t, s, ['stop', 'station'], 'station')
-	t.equal(s.id, barmbek)
+	t.equal(s.id, magdeburgBuckau)
 
 	t.end()
 })
 
 test('radar', async (t) => {
 	const vehicles = await client.radar({
-		north: 53.569,
-		west: 10.022,
-		south: 53.55,
-		east: 10.0436
+		north: 52.148364,
+		west: 11.600826,
+		south: 52.108486,
+		east: 11.651451
 	}, {
 		duration: 5 * 60, when, results: 10
 	})
 
+	const customCfg = Object.assign({}, cfg, {
+		stationCoordsOptional: true, // see #28
+	})
+	const validate = createValidate(customCfg, {})
 	validate(t, vehicles, 'movements', 'vehicles')
+
 	t.end()
 })
