+++ test/e2e/nvv.js
@@ -5,9 +5,9 @@ const tape = require('tape')
 const isRoughlyEqual = require('is-roughly-equal')
 
 const {createWhen} = require('./lib/util')
-const createClient = require('..')
-const nvvProfile = require('../p/nvv')
-const products = require('../p/nvv/products')
+const createClient = require('../..')
+const nvvProfile = require('../../p/nvv')
+const products = require('../../p/nvv/products')
 const createValidate = require('./lib/validate-fptf-with')
 const testJourneysStationToStation = require('./lib/journeys-station-to-station')
 const journeysFailsWithNoProduct = require('./lib/journeys-fails-with-no-product')
