+++ test/index.js
@@ -4,19 +4,5 @@ require('./bvg-journey')
 require('./vbb-departures')
 require('./bvg-radar')
 
-require('./common')
-require('./db')
-require('./vbb')
-require('./bvg')
-require('./oebb')
-require('./insa')
-require('./nahsh')
-require('./cmta')
-require('./nvv')
-require('./sbahn-muenchen')
-require('./saarfahrplan')
-require('./cfl')
-require('./hvv')
-
 require('./throttle')
 require('./retry')
